requirements.txt
django
psycopg2-binary
python-decouple
Pillow
PyJWT